rm tombstone.go
