--- pkg/vm/engine/tae/tables/tombstone.go
@@ -1,62 +1,0 @@
-// Copyright 2021 Matrix Origin
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package tables
-
-import (
-	"context"
-
-	"github.com/matrixorigin/matrixone/pkg/common/mpool"
-	"github.com/matrixorigin/matrixone/pkg/container/types"
-	"github.com/matrixorigin/matrixone/pkg/objectio"
-	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/catalog"
-	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/containers"
-)
-
-func TombstoneRangeScanByObject(
-	ctx context.Context,
-	tableEntry *catalog.TableEntry,
-	objectID objectio.ObjectId,
-	start, end types.TS,
-	mp *mpool.MPool,
-	vpool *containers.VectorPool,
-) (bat *containers.Batch, err error) {
-	it := tableEntry.MakeTombstoneObjectIt(false)
-	for ; it.Valid(); it.Next() {
-		node := it.Get()
-		tombstone := node.GetPayload()
-		tombstone.RLock()
-		skip := tombstone.IsCreatingOrAbortedLocked() || tombstone.HasDropCommittedLocked()
-		tombstone.RUnlock()
-		if skip {
-			continue
-		}
-		visible := tombstone.IsVisibleInRange(start, end)
-		if !visible {
-			continue
-		}
-		if tombstone.HasCommittedPersistedData() {
-			zm := tombstone.GetSortKeyZonemap()
-			if !zm.PrefixEq(objectID[:]) {
-				continue
-			}
-			// TODO: Bloomfilter
-		}
-		err = tombstone.GetObjectData().CollectObjectTombstoneInRange(ctx, start, end, &objectID, &bat, mp, vpool)
-		if err != nil {
-			return nil, err
-		}
-	}
-	return
-}
